VisionCore.GTest: Move synthetic functions to helper.
