+++ tests/vision/core.gtest/CMakeLists.txt
@@ -3,6 +3,7 @@ set(targetName "visionCore.gtest")
 
 # Create executable
 add_executable(${targetName}
+	"${CMAKE_CURRENT_LIST_DIR}/syntheticBoard.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/rectifier.gtest.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/process.gtest.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/boardFinder.gtest.cpp"
